rename am notifications to alert groups
