+++ public/app/plugins/panel/alert-groups/plugin.json
