+++ public/app/plugins/panel/alert-groups/types.ts
@@ -1,4 +1,4 @@
-export interface AMNotificationsOptions {
+export interface AlertGroupPanelOptions {
   labels: string;
   alertmanager: string;
 }
